feat(#3811): reporting twin for parse_value_from_raw_bytes; the short name now asserts full consumption
